requirements.txt
absl-py==2.1.0
accelerate==0.32.1
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.43.1
bleu==0.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.18.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.14.4
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
efficiency==2.0
einops==0.7.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.7.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
interegular==0.3.3
ipykernel==6.29.3
ipython==8.12.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.18
langchain-text-splitters==0.2.2
langsmith==0.1.85
lark==1.1.9
libclang==18.1.1
llama-cloud==0.0.9
llama-index==0.10.55
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.55
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.2.4
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.7
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
modelscope==1.13.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.35.13
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.6
oss2==2.18.4
outlier-utils==0.0.5
OUTLIERS==0.1
outlines==0.0.46
packaging==24.0
pandas==2.2.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
PyPDF4==1.27.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
responses==0.18.0
rich==13.7.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.2
streamlit==1.36.0
striprtf==0.0.26
sympy==1.12
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
text-generation==0.7.0
tf-slim==1.1.0
threadpoolctl==3.4.0
tiktoken==0.7.0
timm==0.9.16
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
triton==2.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
